Fix behavior with single quotes or leading hyphens
I wanted to use this tool to type out shortcuts for commands, but found that due to `xargs` stripping single quotes, I couldn't put quotes in my macros, and if it lead with a hyphen, that would be interpreted as part of the send-keys command. This address both of these issues. It uses a local var, which might not be ideal, but it works for me.